Make inthashmap use size_t.

+++ src/inthashmap.h
@@ -3,21 +3,21 @@
 struct inthashmap
 {
     uint *table;
-    uint tableSize;
-    uint size;
+    size_t tableSize;
+    size_t size;
 };
 
 typedef struct
 {
     const inthashmap *map;
-    uint position;
+    size_t position;
 } inthashmapiterator;
 
-extern nonnull ErrorCode IntHashMapInit(inthashmap *map, uint capacity);
+extern nonnull ErrorCode IntHashMapInit(inthashmap *map, size_t capacity);
 extern nonnull void IntHashMapDispose(inthashmap *map);
 extern nonnull ErrorCode IntHashMapAdd(inthashmap *map, uint key, uint value);
 extern nonnull pure uint IntHashMapGet(const inthashmap *map, uint key);
-extern nonnull pure uint IntHashMapSize(const inthashmap *map);
+extern nonnull pure size_t IntHashMapSize(const inthashmap *map);
 
 extern nonnull void IntHashMapIteratorInit(const inthashmap *map,
                                            inthashmapiterator *iterator);
